+++ main-service/src/main/java/ru/practicum/service/EventService.java
@@ -9,21 +9,21 @@ import java.util.List;
 public interface EventService {
     List<EventFullDto> getAllEventFromAdmin(SearchEventParamsAdmin searchEventParamsAdmin);
 
-    EventFullDto updateEventFromAdmin(Long eventId, UpdateEventAdminRequest inputUpdate);
+    EventFullDto updateEventFromAdmin(Integer eventId, UpdateEventAdminRequest inputUpdate);
 
-    List<EventShortDto> getEventsByUserId(Long userId, Integer from, Integer size);
+    List<EventShortDto> getEventsByUserId(Integer userId, Integer from, Integer size);
 
-    EventFullDto addNewEvent(Long userId, NewEventDto input);
+    EventFullDto addNewEvent(Integer userId, NewEventDto input);
 
-    EventFullDto getEventByUserIdAndEventId(Long userId, Long eventId);
+    EventFullDto getEventByUserIdAndEventId(Integer userId, Integer eventId);
 
-    EventFullDto updateEventByUserIdAndEventId(Long userId, Long eventId, UpdateEventUserRequest inputUpdate);
+    EventFullDto updateEventByUserIdAndEventId(Integer userId, Integer eventId, UpdateEventUserRequest inputUpdate);
 
-    List<ParticipationRequestDto> getAllParticipationRequestsFromEventByOwner(Long userId, Long eventId);
+    List<ParticipationRequestDto> getAllParticipationRequestsFromEventByOwner(Integer userId, Integer eventId);
 
-    EventRequestStatusUpdateResult updateStatusRequest(Long userId, Long eventId, EventRequestStatusUpdateRequest inputUpdate);
+    EventRequestStatusUpdateResult updateStatusRequest(Integer userId, Integer eventId, EventRequestStatusUpdateRequest inputUpdate);
 
     List<EventShortDto> getAllEventFromPublic(SearchEventParams searchEventParams, HttpServletRequest request);
 
-    EventFullDto getEventById(Long eventId, HttpServletRequest request);
+    EventFullDto getEventById(Integer eventId, HttpServletRequest request);
 }
